♻️ Refactor: config 폴더 위치 이동
